Widget: introduce components
Components are subordinates composing the parent widget. There is no
public interface to add or remove them. There are at the end of the
m_subordinates list as they must be drawn last.

Zordering is updated to handle children and components. For instance, a
zorder_top call must not put the child at the end of the list but at the
end of the subset for children.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>