Widget: remove layout() call from the remove_component method
Calling layout() can have nasty side effects when it's done from a
destructor.
For instance, with DialogBox class, the grid component is destructed and
just after the call to the layout method triggers the usage of the grid
component.
This issue was already here before the introduction of subordinates and
components as the layout method is called from the Frame remove method
as well.
Remove it from the remove_component method which is mostly called when
destructing a widget. The layout method can be called explicitely if
needed. Keep it for the Frame remove method as it makes sense to trigger
a layout call when a child is removed.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>