examples/dialog: break shared_ptr reference cycles
Capturing a shared pointer by copy of a widget to set an action for a
child of this widget creates a reference cycle.
The Dialog widget has Button widgets that capture a shared pointer by
copy of the Dialog widget. It prevents the Dialog widget from being
destructed. Capturing the shared pointer by reference is enough to break
the cycle. In this particular case, a weak pointer seems overkilled as
we are sure that Dialog widget exists if an action is triggered by its
children.
This reference cycle hided some bugs with DialogBox that happen only on
destruction.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>