test: add --enable-unittests to configure options
Add --enable-unittests to configure options. It's disabled by default.
The goal of this option is to build the unittests binary and to install
it as it was only built and not installed when using make check.
Moreover make check runs the unit tests, and we don't want the tests to
be run when using build system.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>