PlaneWindow: the show() method bypass the EventLoop
The show() method bypass the EventLoop to display the overlay.

This is needed when switching plane windows, i.e. show window1 and hide
window2. As hide is perfomed instantly, waiting for the EventLoop to
draw and show the plane causes a glitch where we can see the top window.
So do not delayed show to avoid this glitch.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>