Sound: remove classD device fallback from open_alsa_device
The enumeration and the fallback to a hardcoded device makes no sense at
this place. Moreover the device parameter was no longer used, making it
useless and causing a compilation warning.

As it is not easy to figure out the correct name of alsa devices, let's
keep the enumeration feature as a static method of the Sound class.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Fixes: 0a94ca4e8f43 ("sound: enumerate sound devices")