examples/motorcycledash: add new example
Add a new example that uses the SVGDeserial class.

In this example, the background is a PNG picture, the other dynamic graphics
widgets are desgined by SVG, the PNG and SVG file are serialized by the
gfx-convert tool into eraw.h and eraw.bin files. The eraw.h file is included
in the example code, and the eraw.bin file will be loaded when running
the example.

Signed-off-by: Wayne Jia <wayne.jia@microchip.com>
[ludovic.desroches@microchip.com: style.sh script fixes + rework eraw.bin detection]
Tested-by: Ludovic Desroches <ludovic.desroches@microchip.com>