example/camera: invert ctrlwindow and player display
As overlay planes are now displayed immediately and are no longer
waiting for the next iteration of the EventLoop for a redraw, the order
of show() calls may matter.

Player show method has to be called before the one of ctrlwindow to
avoid glitch when computing the layout. As ctrlwindow is bottom aligned,
the player is needed to compute its final position.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>