examples/audioplayer: rework code to catch exceptions
The AudioPlayer widget throws an exception if the is no sound card when
it is constructed.
The audioplayer example doesn't catch this exception. In the case of
linux4microchip distributions, if the board has no sound card, when
selecting the audioplayer example from the launcher, we immediately go
back to the launcher has the app crashes without knowing why.

So rework a bit this example to catch this exception and to display an
error message.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>