ProgressBar: handle the border radius when drawing the progress bar
The border radius was not handled at all when the progress bar was
drawn whereas it was handled for the surrounding widget box.

Hence, the inner progress bar could be drawn outside the rounded box of
the widget.

This patch fixes this issue partically by taking into account the border
radius when drawing the progress bar too.

However, there are some known limitations. Typically, we notice some
blank space around the corners between the widget box and the inner
progress bar.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>