examples/audioplayer: removal of the extra window to display the player
In order to be able to display an error message and to not do too many
changes in the example, an extra window was introduced. The side effect
is an increase of the memory consumption.
Rework the example to remove this extra window. This rework tries to not
introduced too many changes, that's why the shared pointer for m_dial is
not removed.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>