app: change backend logic
At the moment, the libinput backend can't handle a specific input
device. It relies on seat0 which is the default seat an input device is
assigned to by udev. So any input device detected by udev and assigned
to seat0 is handled by the libinput backend.

The libinput backend is always enabled whatever is requested with the
EGT_INPUT_DEVICES environment variable. It means, even if a device is
requested to be handled by a specific backend, if it is detected by udev
and assigned to seat0, it will be handled by the libinput backend as
well. This is probabaly not what is expected.

Let's change the policy. Stop using the libinput backend unconditionally.
Use it only if EGT_INPUT_DEVICES is not set or if it doesn't lead to any
valid input backend. As the libinput backend uses seat0 and is not yet able to
handle a specific device, it's useless to specify it in EGT_INPUT_DEVICES
contrary to what the documentation states. As this feature is going to be
introduced, don't fix the documentation.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>