inputevdev: fix pointer position computation
The pointer position was not well computed. When going through all the
input events received since the last iteration of the event loop, for
performance reasons, the position of the pointer is updated only after
having computed all the pending events. Then it is dispatched.

When a button event is received, it is immadiately dispached. So if it is
surrounded with position events, the pointer position reported is not
correct. Pointer position has to be updated before dispaching the button
event.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>