ScrolledView: fix the 'rect' argument when calling Slider::draw()
ScrolledView::resize_slider() computes the coordinates of the 'm_hslider'
and 'm_vslider' sliders in the subordinate space.

Therefore, in View::draw(), the rectangle argument of the Slider::draw()
method should be relative to this same space too.

Currently, this change is silent because the 'rect' argument of the
Slider::draw() method is not used.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>