ScrolledView: undo the pointer point translation once handled by children
We should restore the pointer point to its original value once
child->handle() has been executed.

Translate the pointer point once for all before entering the 'for' loop
on children, because before that change, the translation could be
applied more than once if event.stop() was not called from
client->handle() and if a further child in the loop also needed to handle
the event.

Also, we want to revert the translation before exiting handle() as this
method doesn't call event.stop() itself(); therefore the event may still
be handled by a further widget.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>