ScrolledView: make all children fit entirely into the "super content area"
When moat() was not zero and when the bottom right point of a child box
was positioned outside of the ScrolledView content area but still inside
its box (for instance, if the child box().bottom_right() point was the
same as the scrolled view box().bottom_right() point), the super_rect()
was computed in such a way that it was imposible to scroll with sliders
or to set the proper offset so the child widget could be drawn entirely.

Change the way super_rect() is computed so all children can be drawn
entirely, even if moat() is not zero.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>