ScrolledView: fix the offset range if a child is placed up/left too much
When to_child(super_rect()).top_left() has any negative coordinate,
which occurs if any child is placed at least partially on the left or on
top of the ScrolledView content area, the offset should be allowed to be
positive, otherwise this child could not be drawn entirely.

However, the offset was forced to be negative or zero. So, this patch
removes this constraint and computes the correct range of values for the
offset.

This range is saved into the new 'm_offset_range' member, so it can be
reused later without computing it again.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>