ScrolledView: fix max_offset() and update_scrollable()
When moat() is not zero, content_area() and box() are no longer the
same.

Then, if all children fit into the content area (for instance if their
fill flags are all set to expand both vertically and horizontally), the
rectangle returned by super_rect() is equal to box().

However, both max_offset() and update_scrollable() compare super_rect()
with content_area(), instead of box().

Therefore, the max offset is non zero and sliders are drawn when they
should not.

Fix this by comparing super_rect() with box().

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>