ScrolledView: fix the draw() method when fill_flags() are not 'solid'
Now the widget box, hence its background if any,  is drawn first at the
beginning of the draw() method, we no longer need to draw another
'solid' background into the canvas.

Instead, now we just zero this canvas before drawing children so
previous drawings are correctly discarded and we no longer force the
cairo compositing operator to CAIRO_OPERATOR_SOURCE, as we want to allow
children to blended if requested.

However, to keep the ScrolledView draw() method backward compatible, we
also default the fill flags to 'solid'.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>