ScrolledView: map directly the offset range to the sliders range
Spare the range translation performed by detail::normalize(); just map
directly the range of values for offset_range.x() to the horizontal
slider and the range of values for offset_range.y() to the vertical
slider.

Also, take advantage of the Slider::value() setter, which already clamps
the new value to garantee it always fit into the range of allowed values.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>