StaticGrid: don't call damage() at the end of reposition()
There is no need to call damage() at the end of reposition().

Indeed, if a child is moved or resized when its box() method is called
from StaticGrid::reposition(), the Widget::move() and Widget::resize()
methods already call damage() with the proper rectangles.

Besides, if no child is either moved or resized, then there is no point
in damaging the StaticGrid since nothing has to be redrawn.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>