inputtslib: use pressure instead of pen_down
Use pressure instead of pen_down to handle touch detection. pen_down
should be driven by BTN_TOUCH, but it seems we can't rely on it. Most of
the tslib examples use the pressure to detect touch events, let's do the
same. Caution, it may make obsolete the plugins dealing with the pen
detection as we rely on the pressure as well.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>