MouseGesture: make distance to travel to enable drop mode configurable
The distance to travel before enabling the drop mode is hard coded. Make
it configurable as this value can be too small and some minor movements
of a finger can be seen as a drag movement.

The static approach is not perfect as there can be several instances of
MouseGesture if several input devices are used, but the API doesn't provide
any access to the m_mouse member of Input.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>