ScrolledView: call Frame::handle() from ScrolledView::handle()
Now that ScrolledView::handle() is very close to Frame::handle(), ie
Widget::handle(), call the inherited method first then do all the event
handling stuff that is specific to the ScrolledView.

Indeed, this way we can share the source code that forwards events to
subordinates and make them handle those events. Hence, we have less
source code to maintain and to modify.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>