Widget: send drag events only to the widget which accepted the drag start
This fixes odd and unwanted behaviors.

For intance, when a 'pointer_drag_start' event is sent to a A widget,
then 'pointer_drag' events leave the A widget and enter a B widget,
the B widget recieves 'pointer_drag' events without having received any
'pointer_drag_start' event first.

However, widgets such as sliders or players from the camera and video
examples, perform some initialization when a 'pointer_drag_start' event
occur, which are required to later handle further 'pointer_drag' events.

Also, when a A widget has B widget as a child and when a 'pointer_drag'
event was sent to A, A always forwards this event to B so at the end
both A and B handle the event. As an example, if a Slider is added as
a child of a ScrolledView, when the user drag the handle of the Slider,
the ScrolledView is scrolled at the same time the handle is moved.

So, this patchs fixes such issues by extending the API so widgets can
now notify the Input instance they caught a 'pointer_drag_start' event,
so:
1 - this very same 'pointer_drag_start' event cannot be caught later by
    another widget.
2 - further 'pointer_drag' and 'pointer_drag_stop' events are only sent
    to the widget which has first caught the 'pointer_drag_event'.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>