examples/calc: fix the size of the EGT logo
The size of the EGT logo is smaller than the native size of the .png
image.

This is because when ImageHolder::do_set_image() is called from the
ImageHolder constructor and resizes the widget first, the widget size is
computed from the current moat, which is zero. Hence, at this time the
widget size is the same as the size of the image.

However, later when the 'elogo' margin is modified, the size of widget
is left unchanged but the image is shrinked, because expanded, taking
into account the new moat.

Then, as we want the size of the image to be its native size, we resize
the 'elogo' after having changed its margin.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>