examples/calc: refactor the layout to take into account the screen size
The layout now adapts itself to the screen size at startup time.

The buttons shape is still a square but its width is proportional to the
screen width.

With landscape screens, we want the button size to remain 60x60 on
800x480 screens, hence buttons are scaled accordingly.

With portrait screens, we want buttons to expand horizontally to fill
the whole screen width.

To simplify the buttons layout, we now add them into a StaticGrid.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>