examples/i18n: set call 'next->zorder_top()' button at startup
'next->zorder_top()' is called only when the 'next' button is clicked.

However, since 'vsizer' is added to 'window' after the 'next' button is,
all labels added to 'vsizer', ie the whole animated text, are drawn on
top of the 'next' button.

Then, when the animated text slides and overlaps then 'next' button',
it prevents the user from clicking on this button.

So, we move that call of 'next->zorder_top()' on startup, once all
widgets have been created and the widgets hierarchy has been completed.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>