Button: change the min size hint for multiline text buttons
If there is text, the min size hint multiplies the size of the text by
three to get the button size. That gives extra space and a good result
when there is only one line of text. When there is several lines of
text, it's too much. In this case, just add few extra pixels.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>