CheckBox,RadioBox: gather command draw source code into Switch
CheckBox::default_draw() and RadioBox::default_draw() are almost the
same; the only difference is how the switch, either a box or a circle,
is drawn.

Therefore, move the common source code into Switch::default_draw() and
make it call a new draw_switch() virtual method to specialize how the
switch is drawn.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>