examples/imagestack: center ImageLabels' image with align flags
All images/imageN.png images have the same 200x400 size.
Also, the PDA5 screen size is 800x480.

To make its image vertically centered, each LauncherItem height was set
to the screen height (480) minus 80, hence 400, so the height of the
internal image.

However, adding the 'center' flag to the image align flags is enough to
be sure the image is vertically centered within each LauncherItem, no
matter the height of this LauncherItem.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>