examples/audioplayer: fix the size of the message dialog
When 'm_message_dialog.resize(this->size() * 0.75)' is executed from the
AudioPlayer constructor, 'this->size()' is {0, 0}, hence the size of
'm_message_dialog' is set to {0, 0}.

Later, when the AudioPlayer instance is added to the top window and
expanded, its layout is computed again. At this step, 'm_message_dialog'
is resized again but this time according to Dialog::min_size_hint().

At the end, the size of 'm_message_dialog' is not 75% of the size of the
AudioPlayer as requested.

This patch fixes that by calling ratio() intead of resize(), so the
'm_message_dialog' is resized correctly when its parent, the AudioPlayer
instance, is resized itself.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>