Slider: track 'pointer_drag' events even outside the slider box
Override the internal_track_drag() method to make it return 'true'.

This way we go back to the previous behavior of sliders, when they
tracked 'pointer_drag' events even when they are outside the slider box,
but still as long as the 'pointer_drag_start' event was inside the
slider box.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>