resolve_path: support for svg icons
The way the icon path was resolved didn't allow to use the icons in the
svg subfolder.
An icon resource is used in this way:
egt::Image("icon:egt_logo_black.png;128").
To resolve the icon subfolder, the size of the icon is concatenated with the
"px" suffix.
Now, if the string provided is not a number, this string is no longer
concatenated with "px", but directly used as the subfolder name.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>