Checkbox,Radiobox: fix theme customization
As the CheckBox and RadioBox no longer have their own draw method, but
rely on the parent method, the Drawer specialization is for Switch.

So when a theme customizes the Drawer for CheckBox, it is no longer
called. Re-introduce the override of the draw method to allow calling
the appropriate specialization of the Drawer.

Fixes: 79b404c20ad8 ("CheckBox,RadioBox: gather command draw source code into Switch")
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>