Slider: fix the prototype of the to_offset() method
The type of the 'value' parameter of the to_offset() method should be 'T'
rather than 'int'.

Indeed, with SliderF widgets, to_offset() used to convert its float
value into an integer, before computing the returned offset.

Since that offset was later used to compute the positon of the slider
handle box, that handle could move not smoothly in some cases.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>