remove gstkmssinkimpl
The gstkmssink is only used for sama5d4 that has a video decoder. As we
no longer release any distribution for sama5d4, and we plan to have a
mainstream solution for this video decoder, drop this sink. Moreover,
this sink is a custom one for the hantro decoder, it's not about the
kmssink provided with GStreamer.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>