VideoWindow: add the gst_custom_pipeline setter
Add the gst_custom_pipeline setter to provide a custom pipeline that
will be used by the GstDecoderImpl.
It's not a good practice to expose implementation, but multimedia
stuff is strongly tied to GStreamer. Moreover, adding this method,
instead of giving direct access to the implementation, prevents
exposing GStreamer to the EGT public interface.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>