GstDecoderImpl: add v4l2src support
The uri is forwarded to the uri parameter of the uridecodebin plugin.
This plugin handles files, not video devices. So if users provide an uri
as file:///dev/video0, the uridecodebin won't be able to deal with it.
Add a detection of /dev/video in the uri. If this pattern is detected,
then switch to the v4l2src plugin to display the video stream.

Take the opportunity to stop using the m_uri variable to store both an uri
that can be used by the uridecodebin plugin and a device node that is used
with the v4l2src plugin. It is less confusing to have a dedicated variable
for uri and one for devices.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>