VideoWindow: remove the create_impl method
Now, there is only one implementation which is the GstDecoderImpl.
Having a create_impl method no longer makes sense, the m_video_impl
member can be directly initialized in the constructor initialization
list.

To not expose the GStreamer types into the EGT public API,
GstDecoderImpl is kept separated from VideoWindow, and VideoWindow uses
an unique pointer on GstDecoderImpl so it can be forward declared.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>