GstDecoderImpl: use its own signals instead of the interface ones
The goal is to have the GstDecoderImpl being the only GStreamer
implementation for all widgets dealing with GStreamer. These widgets
have different signals that are all driven by GStreamer events. So, it
makes sense to move them to GstDecoderImpl and let the widgets choose
the signals there are interested in. They can forward the GstDecoderImpl
signals with the help of signal wrappers.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>