GstDecoderImpl: copy missing features from CameraImpl
In order to merge CamaraImpl in GstDecoderImpl, copy the missing
features from CameraImpl to GstDecoderImpl. This is mostly a raw
copy/paste.
The only changes are about the bus callback and signals.
- The CamaraImpl bus_callback is registered both for the monitor device
bus and the pipeline bus. Its code is the same as the GstDecoderImpl
pipeline bus one plus two events for device detection. So, copy only
the missing code for these events into the GstDecoderImpl
monitor_device_callback.
- GstDecoderImpl has its own signals, so stop using those of the
interface.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>