GstDecoderImpl: rework the create_pipeline_description method
Rework the create_pipeline_description method to handle cameras and to
use it from the start method.
Take this opportunity to improve comments. Options for the appsink
that where used for the CameraWindow but not the VideoWindow have been
skipped as it sounds to have no impact, and it is not explained why
there were added.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Reviewed-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>