GstDecoderImpl: provide a consistent naming for the source
The goal is to provide a consistent naming we can refer to for creating
extra pipeline branches. So there is always an element named source that
corresponds to a usable input.

For the v4l2 source, it requires the addition of a tee element.

The choice of a tee element is based on the fact that replacing the
short notation of the caps filter with the 'normal' notation, in order to
named it, causes issues for the pipeline creation. The GStreamer
documentation states that both notations should be equivalent but it
leads to different pipelines. The one not using the short notation is
not valid, some elements are not linked.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>