Multimedia: rework start/play/media/device methods
Give a clear and coherent responsability to these methods.
Start creates the GStreamer pipeline.
Play calls start if there is no pipeline and set the pipeline to the play
state.
Stop stops the pipeline and destroys it.
Media and device only deals with setting the input and no longer
interfer in pipeline management.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>