Multimedia: encapsulate source elements
Encapsulate the pipeline description of src elements i.e. uridecodebin
and v4l2src. The goal is to simplify the create_pipeline_description
method and to anticipate the needs to merge audio and capture
implementations. GstSrc objects provide the pipeline description for
their part and provide an element named 'source' that can be used in the
overall pipeline description.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>