Multimedia: AudioPlayer makes use of GstDecoderImpl
Most of the code related to GStreamer is similar to GstDecoderImpl.
There are only few features missing. So add the mute feature to
GstDecoderImpl and allow the use of GstDecoderImpl without video sink.

AudioPlayer doesn't set the sync parameter to false for the
autoaudiosink element while GstDecoderImpl does. When set to false, mute
triggers an end of stream event. This parameter comes from GstBaseSink,
but it is not documented. It is not mentionned as well why it is set to
false in GstDecoderImpl so let's keep it to true.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>