GstDecoderImpl: rework get_camera_device_caps()
The get_camera_device_caps method and function are no longer needed. We
let GStreamer choose the camera resolution to use as it is smart enough
to choose the closest resolution to the one requested.

Let's remove the function and keep the method. Even if it is no longer
needed, it is useful to print the resolutions supported. It can help to
choose the window size to avoid software scaling from the GStreamer
pipeline.

Take this opportunity to replace camera with video in the naming of
get_camera_device_caps and get_camera_device_list as the devices are
video/x-raw devices that is wider than camera devices. The naming of
related functions exposed to the API is not updated to not introduce a
breakage.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>