GstDecoderImpl: improve scaling for basic window
For a non-plane window, when a video resize occurs through a scaling
request, only the window is resized. Consequently, GStreamer remains
unaware of this change and continues to provide frames at the previous
size. The necessary scaling to match the new size is handled by Cairo,
resulting in dual scaling operations if there is already one in the
GStreamer pipeline. To resolve this, we can reconfigure the GStreamer
pipeline to manage the scaling for the new window size.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>