ListBox: create an abstract base class for list list boxes
Introduce the ListBoxBase class which is an abstract class that can be
used as a base class to build list boxes widget.
The purpose of this patch is to prepare the addition of a new list box
widget with extra features without breaking the existing one and to
share some common code.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>