ListBoxMulti: introduce a new list box with multi-selection support
Introduce the ListBoxMulti class derived from ListBoxBase. Contrary to
ListBox, it allows multi-selection and an API to easily deal with this
feature.

The selected and deselected methods set the state of the items of the
list. For instance, selected({1, 3, 5}) selects items with index 1, 3, 5
and deselects other items. Select, deselect and toggle methods only
update the state of the requested items.
Events are triggered only when a state change occurs. If an item is
already selected is selected a second time, no event is generated.

An example has been added to test the behavior of the multi-selection
list box.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>