StaticGrid: override add/remove_at methods
The new Frame::add/remove_at methods bypass the StaticGrid logic. To not
corrupt the behavior of StaticGrid, override these methods to stick to
the StaticGrid logic. The position parameter of these methods is
translated in column and row indexes before calling usual StaticGrid
methods.

Update the listboxmulti example that uses a StaticGrid to test these new
methods.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>