GstAppSink: replace the hard-coded RGB565 pixel format with 'm_format'
In the draw() method, the pixel format was hard-coded to RGB565 while it
should be retrieved from the 'm_format' member, which is set in the
GstAppSink constructor, by its GstSink base constructor, from the value
returned by window.format().

Hence, when the window pixel format was something else but RGB565, the
GstAppSink::draw() was broken.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>