PlPlotImpl: set font with cairo_set_scaled_font() in plplot_label()
Cairo documentation states that:

"""
The cairo_select_font_face() function call is part of what the cairo
designers call the "toy" text API. It is convenient for short demos and
simple programs, but it is not expected to be adequete for serious
text-using applications.
"""

Hence, we'd rather use cairo_set_scaled_font() instead.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>