GstAppSink: fix draw() when m_window.plane_window() is true
Change the way the latest gstreamer sample is managed so the draw() method
can always get a valid reference to it, even when the GstAppSink instance
is implemented with a plane window.

Before this patch, the 'm_videosample' member was never set from the
on_new_buffer() method when 'impl->m_window.plance_window()' was true.
Hence, the draw() method could not draw anything.

Especially, when Application::paint_to_file() was executed, the
GstAppSink instance could not be drawn, at least without the
PlaneWindow::paint() overridden method, which is an ugly work-around.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>