Widget: fix draw() to make Widget::paint() work even with plane windows
Widget::draw() can be called from Window::do_draw() or from
Widget::paint().

Widget::draw() normally performs a +point() translation before drawing
the widget's subordinates because their coordinates are relative to the
widget's origin.

However, if the widget had a screen, hence if the widget was either the
TopWindow or some plane window drawn on a dedicated screen overlay,
then:

1 - the widget box origin was forced to (0, 0) in Theme::draw_box()

and

2 -  the +point() translation in Widget::draw() was skipped.

Indeed, Window::do_draw() wants the window to be drawn
at the (0, 0) position of its screen; therefore doing [1] and [2] was
one way to get correct coordinates for both the window's box and its
subordinates/children, when Widget::draw() was called from
Window::do_draw().

On the other hand, with [1] and [2], the box and children coordinates
would have become incorrect if Widget::draw() were called from
Widget::paint() with a plane window at a position different from (0, 0).

As a matter of fact, Widget::paint() also wants the widget to be drawn at
the (0, 0) position in the current space of the Painter instance.

Then to achieve that, Widget::paint() always performs a -point()
translation before calling Widget::draw(), whatever the widget is, but
this translation is incompatible with [1] and [2], if the widget
actually is a plane window.

Then to fix this coordinate issue and to make things more consistent, we
discard [2], meaning the +point() translation in Widget::draw() is now
always performed, but a -point() translation is inserted to
compensate into Window::do_draw(), which is called only on the TopWindow
or plane windows, that is to say on widgets having a screen.
Also [1] is discarded too.

This way, now both Window::do_draw() and Widget::paint() perform a
-point() translation before calling Widget::draw(), which in turn always
performs a +point() translation before drawing the widget's subordinates.

Finally, all overrides of Widget::draw() which don't call directly or
indirectly Widget::draw() but have subordinates anyway (children or
components), should also perform a +point() translation before drawing
these subordinates.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>