Sprite: remove surface() method
The Sprite::surface() method returned a shared_cairo_surface_t.

However it was not used internally in libegt and very unlikely to be
called either from applications linked with libegt.

Hence, we can just remove the Sprite::surface() method, which simplifies
the libegt public API and also eases the abstraction of cairo.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>