Image: remove the unused crop() method
The Image::crop() method is not used at all internally and very unlikely
to be called either from applications linked with libegt.

Besides, it is implemented with the Canvas::copy() method, which itself
calls many cairo functions, which is annoying to abstract cairo.

Therefore, remove the Image::crop() method to simplify the libegt public
API and to ease the abstraction of cairo.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>