Canvas: remove the Canvas class, which is no longer used
The canvas class is made of a cairo_surface_t and a cairo_t created from
this cairo_surface_t.

In the process of abstracting cairo from the EGT public API, remove the
Canvas class.

Anyway, Canvas roles can still be fulfilled with either a Surface instance
alone when only a cairo_surface_t was needed or a Surface instance plus a
Painter instance, created from the Surface, when a cairo_t context was
required.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>