Painter: avoid a cairo translation in Painter::draw(const Image&)
Instead of translating first to point (x, y) then setting the image
surface as the source pattern at (0, 0), directly set the image surface
as the source pattern at (x, y).

Doing so is the same when cairo_paint() is called from Painter::paint(),
because the default extend mode is CAIRO_EXTEND_NONE when a cairo
pattern is created from a surface, as it is the case here, therefore
cairo_paint() draws only within but not outside the source pattern.

At the end, only the actual position of the source surface pattern
matters when paint() is called, whatever the transformation matrix is.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>